chore: rename merge cells to merge cell
