+++ packages/s2-core/__tests__/unit/utils/merge-cell.spec.ts
