+++ packages/s2-core/src/interaction/base-interaction/click/index.ts
@@ -1,4 +1,4 @@
 export { DataCellClick } from './data-cell-click';
-export { MergedCellsClick } from './merged-cells-click';
+export { MergedCellClick } from './merged-cell-click';
 export { RowColumnClick } from './row-column-click';
 export { RowTextClick } from './row-text-click';
