+++ packages/s2-core/src/utils/interaction/index.ts
@@ -1,5 +1,5 @@
 export * from './formatter';
 export * from './hover-event';
-export * from './merge-cells';
+export * from './merge-cell';
 export * from './select-event';
 export * from './state-controller';
